--- a/asav/docker/Dockerfile
+++ b/asav/docker/Dockerfile
@@ -1,0 +1,25 @@
+FROM debian:bullseye
+MAINTAINER Kristian Larsson <kristian@spritelink.net>
+
+ENV DEBIAN_FRONTEND=noninteractive
+
+RUN apt-get update -qy \
+ && apt-get upgrade -qy \
+ && apt-get install -y \
+    bridge-utils \
+    iproute2 \
+    python3-ipy \
+    socat \
+    qemu-kvm \
+    genisoimage \
+ && rm -rf /var/lib/apt/lists/*
+
+ARG VERSION
+ENV VERSION=${VERSION}
+ARG IMAGE
+COPY $IMAGE* /
+COPY *.py /
+
+EXPOSE 22 161/udp 830 5000 10000-10099
+HEALTHCHECK CMD ["/healthcheck.py"]
+ENTRYPOINT ["/launch.py"]