--- a/asav/docker/Dockerfile
+++ b/asav/docker/Dockerfile
@@ -1,4 +1,4 @@
-FROM debian:bullseye
+FROM ubuntu:20.04 
 MAINTAINER Kristian Larsson <kristian@spritelink.net>
 
 ENV DEBIAN_FRONTEND=noninteractive
@@ -11,6 +11,11 @@ RUN apt-get update -qy \
     python3-ipy \
     socat \
     qemu-kvm \
+    tcpdump \
+    ssh \ 
+    inetutils-ping \
+    dnsutils \
+    telnet \
     genisoimage \
  && rm -rf /var/lib/apt/lists/*