Hi all,

This PR is to add support for Cisco ASAv, I've build and tested this with the asav9-18-2.qcow2 image that is available from Cisco CML / VIRL.

I'd like to add it to this project so it hopefully can also be supported in https://github.com/srl-labs/containerlab

Terminal output where you can see the build:
```
root@containerlab:/home/stimmerman/vrnetlab/asav# make docker-image
../makefile-install.include:32: warning: overriding recipe for target 'docker-pre-build'
../makefile.include:18: warning: ignoring old recipe for target 'docker-pre-build'
for IMAGE in asav9-18-2.qcow2; do \
        echo "Making $IMAGE"; \
        make IMAGE=$IMAGE docker-build; \
done
Making asav9-18-2.qcow2
make[1]: Entering directory '/home/stimmerman/vrnetlab/asav'
../makefile-install.include:32: warning: overriding recipe for target 'docker-pre-build'
../makefile.include:18: warning: ignoring old recipe for target 'docker-pre-build'
rm -f docker/*.qcow2* docker/*.tgz* docker/*.vmdk* docker/*.iso
cat cidfile | xargs --no-run-if-empty docker rm -f
Error response from daemon: No such container: 9883b1915fd6c5c922e613750041ec3e289f9fcd56a6e288a49ea811755f2938
rm cidfile
docker tag vr-asav:9-18-2 vr-asav:9-18-2-previous-build
Building docker image using asav9-18-2.qcow2 as vr-asav:9-18-2
cp ../common/* docker/
make IMAGE=$IMAGE docker-build-image-copy
make[2]: Entering directory '/home/stimmerman/vrnetlab/asav'
../makefile-install.include:32: warning: overriding recipe for target 'docker-pre-build'
../makefile.include:18: warning: ignoring old recipe for target 'docker-pre-build'
cp asav9-18-2.qcow2* docker/
make[2]: Leaving directory '/home/stimmerman/vrnetlab/asav'
(cd docker; docker build --build-arg http_proxy= --build-arg https_proxy= --build-arg IMAGE=asav9-18-2.qcow2 --build-arg VERSION=9-18-2 -t vr-asav:9-18-2 .)
[+] Building 4.4s (9/9) FINISHED                                                                                                                                                                                                                                              docker:default
 => [internal] load .dockerignore                                                                                                                                                                                                                                                       0.0s
 => => transferring context: 2B                                                                                                                                                                                                                                                         0.0s
 => [internal] load build definition from Dockerfile                                                                                                                                                                                                                                    0.0s
 => => transferring dockerfile: 531B                                                                                                                                                                                                                                                    0.0s
 => [internal] load metadata for docker.io/library/debian:bullseye                                                                                                                                                                                                                      2.6s
 => [1/4] FROM docker.io/library/debian:bullseye@sha256:1beb7cf458bdfe71b5220cb2069eb45e3fc7eb77a1ccfb169eaebf5f6c4809ab                                                                                                                                                                0.0s
 => [internal] load build context                                                                                                                                                                                                                                                       1.5s
 => => transferring context: 340.35MB                                                                                                                                                                                                                                                   1.5s
 => CACHED [2/4] RUN apt-get update -qy  && apt-get upgrade -qy  && apt-get install -y     bridge-utils     iproute2     python3-ipy     socat     qemu-kvm     genisoimage  && rm -rf /var/lib/apt/lists/*                                                                             0.0s
 => CACHED [3/4] COPY asav9-18-2.qcow2* /                                                                                                                                                                                                                                               0.0s
 => [4/4] COPY *.py /                                                                                                                                                                                                                                                                   0.3s
 => exporting to image                                                                                                                                                                                                                                                                  0.0s
 => => exporting layers                                                                                                                                                                                                                                                                 0.0s
 => => writing image sha256:6f777c6a2b81046b3cbca83a5b2c2ff437621e0b81f5d7baa129cc9045de9240                                                                                                                                                                                            0.0s
 => => naming to docker.io/library/vr-asav:9-18-2                                                                                                                                                                                                                                       0.0s
docker inspect --format '{{.RootFS.Layers}}' vr-asav:9-18-2-previous-build | tr -d '][' | awk '{ $(NF)=""; print }' > built-image-sha-previous
docker inspect --format '{{.RootFS.Layers}}' vr-asav:9-18-2 | tr -d '][' > built-image-sha-current
if [ "$(cat built-image-sha-previous | sed -e 's/[[:space:]]*$//')" = "$(cat built-image-sha-current)" ]; then echo "Previous image is the same as current, retagging!"; \
docker tag vr-asav:9-18-2-previous-build vr-asav:9-18-2 || true; \
else \
echo "Current build differ from previous, running install!"; \
docker run --cidfile cidfile --privileged vr-asav:9-18-2 --trace --install ; \
docker commit --change='ENTRYPOINT ["/launch.py"]' $(cat cidfile) vr-asav:9-18-2; \
docker rm -f $(cat cidfile); \
fi
Current build differ from previous, running install!
2023-08-31 11:56:02,108: vrnetlab   DEBUG    Creating overlay disk image /asav9-18-2-0-overlay.qcow2 for /asav9-18-2.qcow2
2023-08-31 11:56:02,118: vrnetlab   DEBUG    (b"Formatting '/asav9-18-2-0-overlay.qcow2', fmt=qcow2 cluster_size=65536 extended_l2=off compression_type=zlib size=9126805504 backing_file=/asav9-18-2.qcow2 backing_fmt=qcow2 lazy_refcounts=off refcount_bits=16\n", None)
2023-08-31 11:56:02,118: launch     INFO     Installing ASAv
2023-08-31 11:56:02,118: vrnetlab   DEBUG    VM not started; starting!
2023-08-31 11:56:02,118: vrnetlab   INFO     Starting ASAv_vm[0]
2023-08-31 11:56:02,118: vrnetlab   DEBUG    ['qemu-system-x86_64', '-enable-kvm', '-display', 'none', '-machine', 'pc', '-monitor', 'tcp:0.0.0.0:4000,server,nowait', '-m', '2048', '-serial', 'telnet:0.0.0.0:5000,server,nowait', '-drive', 'if=ide,file=/asav9-18-2-0-overlay.qcow2', '-device', 'pci-bridge,chassis_nr=1,id=pci.1', '-device', 'e1000,netdev=p00,mac=52:54:00:59:b4:00', '-netdev', 'user,id=p00,net=10.0.0.0/24,tftp=/tftpboot,hostfwd=udp::2161-10.0.0.15:161,hostfwd=tcp::2830-10.0.0.15:830,hostfwd=tcp::2080-10.0.0.15:80,hostfwd=tcp::2443-10.0.0.15:443,hostfwd=tcp::2022-10.0.0.15:22', '-device', 'e1000,netdev=p01,mac=52:54:00:51:38:01,bus=pci.1,addr=0x2', '-netdev', 'socket,id=p01,listen=:10001', '-device', 'e1000,netdev=p02,mac=52:54:00:83:ba:02,bus=pci.1,addr=0x3', '-netdev', 'socket,id=p02,listen=:10002', '-device', 'e1000,netdev=p03,mac=52:54:00:88:95:03,bus=pci.1,addr=0x4', '-netdev', 'socket,id=p03,listen=:10003', '-device', 'e1000,netdev=p04,mac=52:54:00:4e:d1:04,bus=pci.1,addr=0x5', '-netdev', 'socket,id=p04,listen=:10004', '-device', 'e1000,netdev=p05,mac=52:54:00:85:f9:05,bus=pci.1,addr=0x6', '-netdev', 'socket,id=p05,listen=:10005', '-device', 'e1000,netdev=p06,mac=52:54:00:c3:03:06,bus=pci.1,addr=0x7', '-netdev', 'socket,id=p06,listen=:10006', '-device', 'e1000,netdev=p07,mac=52:54:00:47:94:07,bus=pci.1,addr=0x8', '-netdev', 'socket,id=p07,listen=:10007', '-device', 'e1000,netdev=p08,mac=52:54:00:a5:97:08,bus=pci.1,addr=0x9', '-netdev', 'socket,id=p08,listen=:10008']
2023-08-31 11:56:50,173: launch     TRACE    OUTPUT: loader: Platform type set to default
Platform ASAv
loader: Platform type set
2023-08-31 11:56:52,175: launch     TRACE    OUTPUT:  to default
IO memory blocks requested from bigphys 32bit: 196608
INFO: The system integrator has chosen not to place th
2023-08-31 11:56:54,176: launch     TRACE    OUTPUT: e requested string in the SMBIOS
DEBUG: Could not find string for SMBIOS type *** String ***INFO: SMBIOS indicates Nested 
2023-08-31 11:56:56,179: launch     TRACE    OUTPUT: KVM Environment
early console in extract_kernel
input_data: 0x00000000021c13b1
input_len: 0x0000000000567f93
output: 0x0000000001000000
output_len: 0x0000000001701e20
kernel_total_size: 0x00000000012e8000
trampoline_32bit: 0x000000000009d000
booted via startup_32()
Physical KASLR using RDTSC...
Virtual KASLR using RDTSC...

Decompressing Linux... Parsing ELF... Performing relocations... done.
Booting the kernel.
Memory KASLR using RDTSC...

INIT: version 2.88 bootingaunch     TRACE    OUTPUT: 
Starting udev

2023-08-31 11:57:00,184: launch     TRACE    OUTPUT: Configuring network interfaces... done.
Starting random number generator daemon.

2023-08-31 11:57:02,187: launch     TRACE    OUTPUT: warning: Device /dev/hda has no label.
fsck.fat 4.1 (2017-01-24)
Starting check/repair pass.
Starting verification pass.
/dev/hda1: 14 files, 41415/65246 clusters
dosfsck(/dev/hda1) returned 0
Mounting /dev/hda1
fsck.fat 4.1 (2017-01-24)
Starting check/repair pass.
Starting verification pass.
/dev/hda2: 2 files, 1/2092548 clusters
dosfsck(/dev/hda2) returned 0
Mounting /dev/hda2
warning: Device /dev/hdc has no label.
no cdrom devices found

2023-08-31 11:57:04,189: launch     TRACE    OUTPUT: CAT9k CFG directory mounting
mount: /mnt/cat9k-cfg: special device shared_mount does not exist.
shared_mount Failure
Removing cat9k_cfg

2023-08-31 11:57:06,191: launch     TRACE    OUTPUT: info: Running in kvm virtual environment.
Lina to use serial port /dev/ttyS0 for console IO
Configuring packages on first boot....
 (This may take several minutes. Please do not power off the machine.)
Running postinst /etc/rpm-postinsts/100-sysvinit-inittab...
Running postinst /etc/rpm-postinsts/101-dpdk...
update-rc.d: /etc/init.d/run-postinsts exists during rc.d purge (continuing)
 Removing any system startup links for run-postinsts ...
  /etc/rcS.d/S99run-postinsts


Loading...
The digital signature of the running image verified successfully
lina_init_env: memif is not enabled.
System Cores 1 Nodes 1 Max Cores 32

total_reserved_mem = 0 

total_heapcache_mem = 0 
total mem 2064969728 system 2098524160 kernel 12596633 image 0
new 1247066727 old 2064969728 reserve 0 priv new 1259663360 priv old 1259663360
Processor memory:   2147483648
M_MMAP_THRESHOLD 65536, M_MMAP_MAX 32768
POST started...
POST finished, result is 0 (hint: 1 means it failed)

Cisco Adaptive Security Appliance Software Version 9.18(2)

Compiled on Mon 08-Aug-22 15:58 GMT by builders
Failed to read security parameters - base 0xfff00000 offset 0x400 buf_size 20
secstore_buf_fill: Error reading secure store -  buffer 0x46294c40, size 0x14 tag 3 id 0
ASA: Platform type set to default. secstore rcode 1
Failed to read security parameters - base 0x0 offset 0x400 buf_size 20
secstore_buf_fill: Error reading secure store -  buffer 0x46294770, size 0x14 tag 4 id 0
CPU mode is not set and it may have impact on performance. Please check the ASAv installation document for appropriate cpu mode.

2023-08-31 11:57:16,202: launch     TRACE    OUTPUT: 
Total NICs found: 10
en_vtun rev00 Backplane Tap Interface     @ index 09 MAC: 0000.0100.0001
Counter ID 'TLS13_DOWNSTREAM_CLIENT_CERTIFICATE_VERIFY' is too long must be 40 characters or less
WARNING: Attribute already exists in the dictionary.

INFO: Unable to read firewall mode from flash
       Writing default firewall mode (single) to flash

INFO: Unable to read cluster interface-mode from flash
        Writing default mode "None" to flash
Error: Platform type has not been configured.
Use software crypto.

2023-08-31 11:57:18,204: launch     TRACE    OUTPUT: 
Cisco Adaptive Security Appliance Software Version 9.18(2) 

  ****************************** Warning *******************************
  This product contains cryptographic features and is
  subject to United States and local country laws
  governing, import, export, transfer, and use.
  Delivery of Cisco cryptographic products does not
  imply third-party authority to import, export,
  distribute, or use encryption. Importers, exporters,
  distributors and users are responsible for compliance
  with U.S. and local country laws. By using this
  product you agree to comply with applicable laws and
  regulations. If you are unable to comply with U.S.
  and local laws, return the enclosed items immediately.

  A summary of U.S. laws governing Cisco cryptographic
  products may be found at:
  http://www.cisco.com/wwl/export/crypto/tool/stqrg.html

  If you require further assistance please contact us by
  sending email to export@cisco.com.
  ******************************* Warning *******************************
Cisco Adaptive Security Appliance Software, version 9.18
Copyright (c) 1996-2022 by Cisco Systems, Inc.
For licenses and notices for open source software used in this product, please visit
http://www.cisco.com/go/asa-opensource

                Restricted Rights Legend
Use, duplication, or disclosure by the Government is
subject to restrictions as set forth in subparagraph
(c) of the Commercial Computer Software - Restricted
Rights clause at FAR sec. 52.227-19 and subparagraph
(c) (1) (ii) of the Rights in Technical Data and Computer
Software clause at DFARS sec. 252.227-7013.

                Cisco Systems, Inc.
                170 West Tasman Drive
                San Jose, California 95134-1706

Error: Platform type has not been configured.

 Successfully discovered platform.  Rebooting to apply the platform type.

2023-08-31 11:57:20,206: launch     TRACE    OUTPUT: Process shutdown finished

2023-08-31 11:57:22,209: launch     TRACE    OUTPUT: Rebooting... (status 0x9)

2023-08-31 11:57:30,219: launch     TRACE    OUTPUT: ..
INIT: Switching to runlevel: 6
INIT: Sending processes the TERM signal

2023-08-31 11:57:32,222: launch     TRACE    OUTPUT: Stopping OpenBSD Secure Shell server: sshd
no /usr/sbin/sshd found; none killed
Stopping Advanced Configuration and Power Interface daemon: no /usr/sbin/acpid found; none killed
acpid.
Stopping random number generator daemon.
Deconfiguring network interfaces... done.
Sending all processes the TERM signal...

2023-08-31 11:57:36,226: launch     TRACE    OUTPUT: Sending all processes the KILL signal...
Deactivating swap...
Unmounting local filesystems...
Rebooting... 
2023-08-31 11:58:28,280: launch     TRACE    OUTPUT: Platform ASAv
IO memor
2023-08-31 11:58:30,282: launch     TRACE    OUTPUT: y blocks requested from bigphys 32bit: 196608
INFO: The system integrator has chosen not to place the requested string in t
2023-08-31 11:58:32,283: launch     TRACE    OUTPUT: he SMBIOS
DEBUG: Could not find string for SMBIOS type *** String ***INFO: SMBIOS indicates Nested KVM Environment

2023-08-31 11:58:34,285: launch     TRACE    OUTPUT: early console in extract_kernel
input_data: 0x00000000021c13b1
input_len: 0x0000000000567f93
output: 0x0000000001000000
output_len: 0x0000000001701e20
kernel_total_size: 0x00000000012e8000
trampoline_32bit: 0x000000000009d000
booted via startup_32()
Physical KASLR using RDTSC...
Virtual KASLR using RDTSC...

Decompressing Linux... Parsing ELF... Performing relocations... done.
Booting the kernel.
Memory KASLR using RDTSC...

INIT: version 2.88 bootingaunch     TRACE    OUTPUT: 
Starting udev

2023-08-31 11:58:40,292: launch     TRACE    OUTPUT: Configuring network interfaces... done.
Starting random number generator daemon.
warning: Device /dev/hda has no label.
fsck.fat 4.1 (2017-01-24)

2023-08-31 11:58:42,294: launch     TRACE    OUTPUT: Starting check/repair pass.
Starting verification pass.
/dev/hda1: 14 files, 41415/65246 clusters
dosfsck(/dev/hda1) returned 0
Mounting /dev/hda1
fsck.fat 4.1 (2017-01-24)
Starting check/repair pass.
Starting verification pass.
/dev/hda2: 21 files, 57/2092548 clusters
dosfsck(/dev/hda2) returned 0
Mounting /dev/hda2
warning: Device /dev/hdc has no label.
no cdrom devices found

2023-08-31 11:58:50,302: launch     TRACE    OUTPUT: Info: Encrypted disk file system created & mounted successfully


2023-08-31 11:58:52,304: launch     TRACE    OUTPUT: CAT9k CFG directory mounting
mount: /mnt/cat9k-cfg: special device shared_mount does not exist.
shared_mount Failure
Removing cat9k_cfg

2023-08-31 11:58:54,307: launch     TRACE    OUTPUT: info: Running in kvm virtual environment.
Lina to use serial port /dev/ttyS0 for console IO
Configuring packages on first boot....
 (This may take several minutes. Please do not power off the machine.)
Running postinst /etc/rpm-postinsts/100-sysvinit-inittab...
Running postinst /etc/rpm-postinsts/101-dpdk...
update-rc.d: /etc/init.d/run-postinsts exists during rc.d purge (continuing)
 Removing any system startup links for run-postinsts ...
  /etc/rcS.d/S99run-postinsts


Loading...
The digital signature of the running image verified successfully
lina_init_env: memif is not enabled.
System Cores 1 Nodes 1 Max Cores 32

total_reserved_mem = 0 

total_heapcache_mem = 0 
total mem 2064969728 system 2098524160 kernel 12596633 image 0
new 1247066727 old 2064969728 reserve 0 priv new 1259663360 priv old 1259663360
Processor memory:   2147483648
M_MMAP_THRESHOLD 65536, M_MMAP_MAX 32768
POST started...
POST finished, result is 0 (hint: 1 means it failed)

Cisco Adaptive Security Appliance Software Version 9.18(2)

Compiled on Mon 08-Aug-22 15:58 GMT by builders
CPU mode is not set and it may have impact on performance. Please check the ASAv installation document for appropriate cpu mode.

2023-08-31 11:59:10,323: launch     TRACE    OUTPUT: 
Total NICs found: 10
en_vtun rev00 Backplane Tap Interface     @ index 09 MAC: 0000.0100.0001
Counter ID 'TLS13_DOWNSTREAM_CLIENT_CERTIFICATE_VERIFY' is too long must be 40 characters or less
WARNING: Attribute already exists in the dictionary.
Use software crypto.

Cisco Adaptive Security Appliance Software Version 9.18(2) 

…